requirements.txt
git+https://github.com/cluic/wxauto.git
wechatpy
flask
pycryptodome
google-generativeai
pyyaml
requests
